# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Listing {
  id: number;
  price: number;
  address: string;
  city: string;
  state: string;
  bedrooms: number;
  bathrooms: number;
  sqft: number;
  imageUrl: string;
  propertyType: string;
  description: string;
  saved: boolean;
}

export type SwipeRequestDirection =
  (typeof SwipeRequestDirection)[keyof typeof SwipeRequestDirection];

export const SwipeRequestDirection = {
  left: "left",
  right: "right",
} as const;

export interface SwipeRequest {
  direction: SwipeRequestDirection;
}

export interface SwipeResponse {
  success: boolean;
  saved: boolean;
}