# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /listings/{id}/swipe:
    post:
      operationId: swipeListing
      tags: [listings]
      summary: Swipe on a listing
      description: Record a swipe action (save or reject) on a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SwipeRequest"
      responses:
        "200":
          description: Swipe recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SwipeResponse"
...
components:
  schemas:
...
    SwipeRequest:
      type: object
      properties:
        direction:
          type: string
          enum: [left, right]
      required:
        - direction
    SwipeResponse:
      type: object
      properties:
        success:
          type: boolean
        saved:
          type: boolean
      required:
        - success
        - saved